+++ R/arc-select.R
@@ -623,6 +623,12 @@ validate_page_size <- function(
 # Protocol Buffer helpers ------------------------------------------------
 
 supports_pbf <- function(x, arg = rlang::caller_arg(x), call = rlang::caller_call()) {
+  check_inherits_any(
+    x,
+    class = c("FeatureLayer", "Table", "ImageServer"),
+    arg = arg,
+    call = call
+  )
   # verify that x is an layer
   # FIXME: This check makes arc_select error on ImageServer inputs
   check_inherits_any(